Create runners only if there is no runners (for demo purpose)

+++ app/src/main/java/utbm/lepysidawy/code_p25/MainActivity.java
@@ -28,44 +28,52 @@ public class MainActivity extends AppCompatActivity {
         AppDatabase db = AppDatabase.getInstance(this);
         Log.d("App","Running");
 
-        List<ParticipateRace> participations = db.participateRaceDAO().getAll();
-        for(ParticipateRace p : participations){
-            db.participateRaceDAO().delete(p);
-        }
 
         List<Runner> runners = db.runnerDAO().getAll();
-        for(Runner r : runners){
-            db.runnerDAO().delete(r);
+        if(runners.size() == 0)
+        {
+            List<ParticipateRace> participations = db.participateRaceDAO().getAll();
+            for(ParticipateRace p : participations){
+                db.participateRaceDAO().delete(p);
+            }
+            for(Runner r : runners){
+                db.runnerDAO().delete(r);
+            }
+
+            List<Race> races = db.raceDAO().getAll();
+            for(Race r : races){
+                db.raceDAO().delete(r);
+            }
+
+            Runner paulMcCartney = new Runner("Paul","McCartney",80);
+            Runner johnLennon = new Runner("John","Lennon",40);
+            Runner ringoStarr = new Runner("Ringo","Starr",90);
+            Runner georgeHarrison = new Runner("George","Harrison",95);
+            Runner mickJagger = new Runner("Mick","Jagger",70);
+            Runner keithRichards = new Runner("Keith","Richards",90);
+            Runner robertPlant = new Runner("Robert","Plant",75);
+            Runner markKnopfler = new Runner("Mark","Knopfler",80);
+            Runner davidGilmour = new Runner("David","Gilmour",78);
+            Runner liamGallagher = new Runner("Liam","Gallagher",62);
+            Runner noelGallagher = new Runner("Noël","Gallagher",33);
+            Runner thomYorke = new Runner("Thom","Yorke",98);
+            Runner davidBowie = new Runner("David","Bowie",53);
+            Runner louReed = new Runner("Lou","Reed",47);
+            Runner brianWilson = new Runner("Brian","Wilson",22);
+
+
+
+
+            long[] runnerIds = db.runnerDAO().insertAll(paulMcCartney,johnLennon,ringoStarr, georgeHarrison, mickJagger, keithRichards, robertPlant, markKnopfler, davidGilmour,
+                    liamGallagher, noelGallagher, thomYorke, davidBowie, louReed, brianWilson);
         }
 
-        List<Race> races = db.raceDAO().getAll();
-        for(Race r : races){
-            db.raceDAO().delete(r);
-        }
 
-        Runner paulMcCartney = new Runner("Paul","McCartney",80);
-        Runner johnLennon = new Runner("John","Lennon",40);
-        Runner ringoStarr = new Runner("Ringo","Starr",90);
-        Runner georgeHarrison = new Runner("George","Harrison",95);
-        Runner mickJagger = new Runner("Mick","Jagger",70);
-        Runner keithRichards = new Runner("Keith","Richards",90);
-        Runner robertPlant = new Runner("Robert","Plant",75);
-        Runner markKnopfler = new Runner("Mark","Knopfler",80);
-        Runner davidGilmour = new Runner("David","Gilmour",78);
-        Runner liamGallagher = new Runner("Liam","Gallagher",62);
-        Runner noelGallagher = new Runner("Noël","Gallagher",33);
-        Runner thomYorke = new Runner("Thom","Yorke",98);
-        Runner davidBowie = new Runner("David","Bowie",53);
-        Runner louReed = new Runner("Lou","Reed",47);
-        Runner brianWilson = new Runner("Brian","Wilson",22);
 
         /**Random rd = new Random();
-        for(int i = 0; i<30; i++){
-            db.runnerDAO().insertAll(new Runner("A" + i, "B"+ i,rd.nextInt(100)));
-        }**/
-
-        long[] runnerIds = db.runnerDAO().insertAll(paulMcCartney,johnLennon,ringoStarr, georgeHarrison, mickJagger, keithRichards, robertPlant, markKnopfler, davidGilmour,
-                liamGallagher, noelGallagher, thomYorke, davidBowie, louReed, brianWilson);
+         for(int i = 0; i<30; i++){
+         db.runnerDAO().insertAll(new Runner("A" + i, "B"+ i,rd.nextInt(100)));
+         }**/
     }
 
     public void onRaceCreationClick(View view) {
